2007 version of the presentation
In 2007 I presented this during Moita Lan Party 2007.

I think I did this presentation later, but I sort of suck at keeping up my
files organized, and this is the latest version of the presentation I can find.

There is a video recording of the delivery of this particular presentation at
https://www.youtube.com/watch?v=7v7UiqmeSTw , if you're interested.

FYI, the presentation was delivered in 1h17m (Q&A happened during the actual
presentation).